Prevent loading images that would require too much memory.
For most image formats, the header specifies the width/height.
PHP allocates an image object from that size, even if the actual
image data is much smaller. This image object size is not limited
by the limit configured in PHP.

The memory limit can be configured through setting "memory_limit"
in app "preview" and defaults to 128 MBytes which should be enough
for most images without filling up all memory.

Signed-off-by: Joachim Bauch <bauch@struktur.de>